refactor: 軽微なリファクタ
